finish up db rollback checks
- move catch_db_error to utils
- tidy catch/propagate errors in prepare, get_current_user
